eos-shard-writer: Fix where we write blob entries
In commit 6633a84, I removed data_offs, but forgot to adjust the math of
where we stored the files to adjust for the header_size field, which is
part of our simple GVariant wrapper.

As such, if we were close enough to our alignment boundary, the header
would stomp on file contents, giving incorrect contents.

[endlessm/eos-sdk#3210]