eos-shard-shard-file: Fix find_record_by_raw_name
Ever since the GVariant conversion in e291724, we store
EosShardRecord *s in our internal array, not EosShardRecord entries
directly. The bsearch API returns a pointer into our array, so we
effectively have an EosShardRecord **, which we have to dereference
before we can pull out the EosShardRecord *.

We also need to make sure we ref it before we return it to gjs, so it
won't free the record accidentally.