shard-file: Improve performance loading large files
Loading the encyclopedia shard header is synchronously and involves a
lot of allocations up-front. We used to perform a lot better by only
doing the allocations when requested, so let's fall back to that
behavior.

The list_records method became a bit uglier and slower during this
process, but list_records should be considered a debugging method, so
that's considered an OK tradeoff. If the performance of list_records
becomes an issue, we should look into fixing that separately.