format: Allow blobs with no contents
Some records in shards might not have associated data, only metadata. We
could also imagine some records that might not want associated metadata.
Let's provide a way to mark a blob as saying it does not have any
contents, and respect it everywhere we load contents or return a stream.