build: Use versioned library names and sonames
This adds the API version to the end of the library name, so that e.g.
libeos-shard-0.so can be parallel installed with libeos-shard-1.so.

This also adds a libtool version number (A.K.A. soname) to the end of the
filename: libeos-shard-0.so.0.0.0. This number is automatically
calculated using the package version number and an "interface age", using
a scheme originally introduced by Emmanuele.

PS. There's an alternative approach to automatically calculating the
soname, explained here:
https://developer.gnome.org/programming-guidelines/unstable/parallel-installability.html.en#version-numbers
This separates the soname from the package version. I do like the
cleanliness of that approach but I chose instead to do the same thing
that we do in our other packages.

[endlessm/eos-sdk#3150]