blob: Remove content type field
Accurate content type reading inside the shard hasn't given us much --
the engine still uses contentType stuffed inside the JSON. It's just
caused us more pain at build time, to be honest, with it in both places.

For the proposed Shard V3 format, we're not going to use it, so remove
it now.