Add a new ShardV2 format
The new V2 format completely removes GVariant-based serialization --
partly because it's slow for our use case, partly because it's not
useful for the record inserting and shard merging tricks we want to
eventually do.

The new format has a few new properties. The main one is that we now
support named blobs per shard. The API hasn't changed yet -- we still
support the two "classic" metadata/data pairs, but it's expected that if
we want to reference additional blobs from inside a record (think:
CSS/JS files for articles, separate thumbnail images for videos), we can
do it with this flat FST.

The expectation is that a record can do relative lookups with something
like this.

Additionally, while not represented in the API yet, it is OK and allowed
for blobs to be not tied to any record, which we might do for Xapian
DBs. The "new" expectation is that records are user-visible (think: has
a card).

The format is also designed to be extendable and forward-compatible with
the addition of capability bits and flags fields to most major data
structures.

This also updates the writer to use the new format as well, removing all
forms of legacy V1 serialization. If V1 serialization is wanted, we can
attempt to bring it back with a separate class.