shard-file: Don't set error state on cancel
If an eos_shard_file_init_async() operation is cancelled, then we should
not store the G_IO_ERROR_CANCELLED in the shard file's internal error
state. Instead, the shard file should still be inited, and any other
pending init operations should succeed.

Per the documentation of g_task_run_in_thread() and
g_task_set_return_on_cancel(), the task function in the other thread will
still run to completion even if the operation is cancelled, unless we set
return-on-cancel to TRUE (which we don't.) So the shard file will still
be initialized, and we only have to make sure we return a
G_IO_ERROR_CANCELLED on any operations that were cancelled.

https://phabricator.endlessm.com/T10595