shard-writer-v2: Add the skeleton of a new multi-threaded writer impl
This isn't actually multithreaded yet, but all the pieces are there.

The algorithm is commented within the implementation, which should
hopefully be sufficient, but basically, it uses several Linux-specific
features to "punch a hole" in the file, allowing us to calculate offsets
for the full uncompressed file, and then remove pages afterwards so they
don't take up space.