shard-writer-v2: Use an off_t for these offsets
The values that we pass to pwrite should be off_t's, since an off_t is a
file offset. On most (all?) systems we care about, this will be a 64-bit
integer since we ask for large file support, but C types are weird and I
would just prefer to use the "right" ones.