eos-shard-writer-v2: Compress to a separate file first
We have a problem of wanting to compress multiple files into a shard in
parallel. But doing it in parallel means that, since we don't know how
big a file is going to be after compression, it's hard to pack them back
to back. We originally attempted the "fallocate" strategy where we
allocate space for the entire file, and then copy that in, and then use
a Linux-specific API call to "collapse" blocks that ended up being
unused. But since the block size of a traditional ext4 mount is 4K, that
ended up wasting too much space. In a 10k encyclopedia shard, we found
that the empty space ended up being 1GB out of the total 2.5GB.

We could try to "use" that wasted space by allowing other uncompressed
files to fill that gap once written, but that requires maintaining a
sorted array of gaps and blobs in the file, and I got halfway through it
before wondering if we could just rethink the situation.

So I came up with this idea: just compress to a different file. This
means that we're spending 3x the IO per blob: first download into a
tempfile, then compress it into another tempfile, then copy that
tempfile into the shard, but it gets us back to our 20 byte alignment
instead of the 4K one, removes the Linux-specific API calls, and other
jazz like that.

There is a future idea: do the compression ahead-of-time at artifactify
time. This works well with the tempfile approach, since we just have to
introduce a new flag for "don't recompress this, but mark it as zlib
compressed".