eos-shard-writer-v2: Be more careful with fd's
The GFileInputStream "owns" its own fd, which it will close at object
destruction time. But we also have our own FD which we close for the
tmpfile case. This "double-close" doesn't immediately show up in our
test suite, but would clearly enable a scenario where we are closing
another thread's fd's.

Since it's impossible to get the GInputStream to not close its fd, just
take our own dup of the fd, and let it close its own fd as soon as
possible.

Also, check the result of the close(1) syscall, just so we don't make
this mistake in the future.