shard-writer-v2: Deduplicate blob data based on checksum
In practice, we've noticed that shards tend to contain duplicate
data. Let's reduce the amount of data stored in a shard by deduplicating
blob contents themselves.

Since we already calculate blob checksums to place in headers, use that
same checksum to skip writing data we already know is inside the shard.

This also now places blob headers to be separate from blob data, which
is required since blobs might have the same data but different flags or
different names for each blob. We now have a rough structure of:

  magic
  shard header
  blob contents
  blob headers
  record/blob tables
  record headers
  string constant table

This has two implications on performance: first, client read performance
might be influenced because separating blob headers from blob contents
adds a tiny bit of indirection that might cause another disk block
read. However, the format is already quite inefficient in terms of disk
reads. If required, we could make a few adjustments to increase page and
cache efficiency while retaining the format.

Second, the writer performance might be influenced because we read the
same file in twice: once to compute the checksum, twice to copy it to
the shard file, whereas before this was done in the same pass. Since
these are "right next to each other", I expect that any reads would
remain in the kernel's block cache, and thus, I don't expect a lot of
performance loss here.

This change should be 100% compatible with old shard libraries -- we
only changed the writer, and not the format nor reader.

https://phabricator.endlessm.com/T13671